bloodwork-dashboard/backend/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
supabase==1.0.3
python-dotenv==1.0.0
elevenlabs==0.2.24
pydantic>=1.10.0,<2.0.0 